Move rockets to components folder.
